% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StabilityAnalysis.R
\name{boots.data}
\alias{boots.data}
\title{Generate a bootstrap sample and divide data to train and test sets}
\usage{
boots.data(data, train.portion)
}
\arguments{
\item{data}{\code{featureSelection} would give the input. Basically, a data frame consisting numTS matrix and outcome score (in the last column).}

\item{train.portion}{Portion of train set.}
}
\value{
A list consisting 1) train set predictor matrix, 2) train set outcome matrix, 3) test set predictor matrix, 4) test set outcome matrix, 5) number of features.
}
\description{
\code{boots.data} is called in \code{featureSelection} function to generate a bootstrap sample and divide data into train and test set.
}
\details{
This is a function called in \code{featureSelection} function. In each iterative run of model training, it generates a bootstrap sample and splits whole dataset into train and test set, so that model trained in train set can be evaluated in the completely unseen test set. It allows us to do feature selection as well as evaluate model performance.
}
\examples{
div.data  <- divide.data(df.numTS, train.portion = 0.75)

}
