% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTrainingNestedCV.R
\name{divide.data}
\alias{divide.data}
\title{Divide data to train and test (outer loop)}
\usage{
divide.data(data, train.portion)
}
\arguments{
\item{data}{\code{trainModel} would give the input. Basically, the input is a data frame consisting predictor matrix and outcome score ( the last column). One of two data frames included in the list object returned by \code{process_data}.}

\item{train.portion}{Portion of train set.}
}
\value{
A list consisting 1) train set predictor matrix, 2) train set outcome matrix, 3) test set predictor matrix, 4) test set outcome matrix, 5) number of features.
}
\description{
\code{divide.data} is called in \code{trainModel} function to divide data.
}
\details{
This is a function called in \code{trainModel} function. In each iterative run of model training, it splits whole dataset into train and test set, so that model trained in train set can be evaluated in the completely unseen test set.
}
\examples{
div.data  <- divide.data(dataList$df, train.portion = 0.75)

}
