% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StabilityAnalysis.R
\name{featureSelection}
\alias{featureSelection}
\title{Feature selection using Elastic net regression conjugated with bootstrapping}
\usage{
featureSelection(Data, num_simulation, trainportion = 0.75)
}
\arguments{
\item{Data}{A data frame consisting predictor numTS matrix and outcome score (in the last column). Please check our example data (Example_TCGA_numTS.txt and Example_Seo_numTS.txt) to make sure there is no formatting issue.}

\item{num_simulation}{Desired number of bootstrap samples.}

\item{trainportion}{Desired portion of train set. Default to 0.75.}
}
\value{
A list containing 1) table of miRNA with the selected frequency and direction of association. 2) R-square estimated in the test set. 3) RMSE estimated in the test set.
}
\description{
\code{featureSelection} generates bootstrap sample and splits data into train set and test set. It trains model using elastic net regression and evaluate the performance by calculating R square and RMSE on test set.
}
\details{
\code{trainModel} train predictive models using elastic net regression conjugated with bootstrapping. It conducts elastic net regression conjugated with bootstrapping to select stable sets of predictive miRNAs. It first creates \code{num_simulation} bootstrap samples from the original dataset by resampling with replacement. Then, with each bootstrap sample, it selects miRNAs using elastic net regression. This step would generate \code{num_simulation} sets of selected miRNAs from \code{num_simulation} bootstrap samples. Then, we calculated the frequency of being selected for each miRNA to indicate its predictive power.
}
\examples{
Example_TCGA_numTS <- read.delim("./data/Example_TCGA_numTS.txt")
resultExample <- featureSelection(Data = Example_TCGA_numTS, num_simulation = 200)


}
