% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{process_data}
\alias{process_data}
\title{Read in output of PRIMATA-APA, process the data for model training}
\usage{
process_data(numTSdir, EXPRdir, hallmark)
}
\arguments{
\item{numTSdir}{Path to the numTS table output from PRIMATA-APA. Please check our example data (Example_numTS.txt) to make sure there is no formatting issue.}

\item{hallmark}{User provided outcome features to be predicted. Please check our example data (Example_hallmark.txt) to make sure there is no formatting issue.}

\item{Exprdir}{Path to the miR expression table output from PRIMATA-APA. Please check our example data (Example_EXPR.txt) to make sure there is no formatting issue.}
}
\value{
The return will be a list of two data frames consisting numTS predictors with outcome score and expression predictors with outcome score.
}
\description{
Read in the output of PRIMATA-APA, ensure the miRNA expression and numTS features are matched, take log transformation for numTS features, merge outcome scores with predictor matrices.
}
\details{
\code{process_data} will read in the output of PRIMATA-APA programs (provided on the github) and process the data for predictive model training.
}
\examples{
dataList  <- process_data(numTSdir="Path/to/numTSfile", EXPRdir="Path/to/Exprfile", hallmark="Path/to/outcomeScorefile")


}
